LayoutTests/js/eval-cache-scoped-lookup-expected.txt

Test to make sure that scoped variable access through an eval produces the correct results

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS first() is 'first'
PASS second() is 'second'
PASS third() is 'third'
PASS fourth() is 'fourth'
PASS fifth() is 'fifth'
PASS sixth() is 'sixth'
PASS seventh() is 'seventh'
PASS eighth() is 'eighth'
PASS nineth() is 'nineth'
PASS tenth() is 'tenth'
PASS eleventh() is 'eleventh'
PASS successfullyParsed is true

TEST COMPLETE
